+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.1 (Jul 14, 2019)
+
+- Fixed general stylings, no more fixed positioning
+- Moved capture button out from camera component
+
 # 1.0.0 (Jul 11, 2019)
 
 - Fixed general stylings